+++ test/performance/run-cruds-benchmark.ts
@@ -1,9 +1,0 @@
-import { runCrudsBenchmark } from './cruds-benchmark'
-
-(() => {
-  const result = runCrudsBenchmark()
-  console.table(result.map(e => ({
-    name: e.name,
-    totalTime: e.totalTime
-  })))
-})()